requirements_rag_minimal.txt
# Minimal requirements - install in order
# Run: pip install -r requirements_rag_minimal.txt

# Core
streamlit>=1.39.0
anthropic>=0.18.0
pandas>=2.2.3
openpyxl>=3.1.5
python-dotenv>=1.0.0
plotly>=5.24.0
numpy>=1.26.4
scipy>=1.14.0

# Database
duckdb>=0.10.0
sqlalchemy>=2.0.25

# Vector Store (optional - can skip if issues)
# chromadb>=0.4.22
# sentence-transformers>=2.3.1

# Optional
# streamlit-chat>=0.1.1
# redis>=5.0.0
